# Conflicts:
#	pom.xml

+++ mzdb-access/src/main/java/fr/profi/mzdb/io/reader/cache/MzDbEntityCache.java
@@ -33,7 +33,7 @@ public class MzDbEntityCache {
 	
 	protected Map<Integer, ArrayList<Long>> spectrumIdsByTimeIndex = null;
 
-	protected Map<Integer, DataEncoding> dataEncodingById = null;
+	protected Map<Long, DataEncoding> dataEncodingById = null;
 
 	protected Map<Long, DataEncoding> dataEncodingBySpectrumId = null;
